Improvements
This version removes the lag between sensing an object and vibrating the bands. The lag is resolved by manually triggering each sensor rather than depending on the automatic triggering offered by the sensors. A rolling average is also applied to increase sensitivity to objects at a close range that enter the FOV suddenly.